refactor(parser): extract parse_layer_opts into a separate function
